pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scitex-orochi"
version = "0.14.0"
description = "Agent Communication Hub for the SciTeX ecosystem"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [
    { name = "Yusuke Watanabe", email = "ywatanabe@scitex.ai" },
]
keywords = ["agent", "communication", "websocket", "hub", "scitex"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "django>=5.1",
    "channels>=4.0",
    "daphne>=4.0",
    "whitenoise>=6.0",
    "django-allauth[socialaccount]>=65.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "aiosqlite>=0.20",
    "websockets>=12.0",
    "pywebpush>=1.14",
    "py-vapid>=1.9",
    # todo#47 — web-terminal SSH backend. Pure-Python asyncio SSH client
    # so the terminal consumer can stream into fleet hosts without
    # shelling out to the openssh binary.
    "asyncssh>=2.14",
]

[project.optional-dependencies]
mcp = ["fastmcp>=2.0"]
agent-container = ["scitex-agent-container>=0.8.0"]
telegram = ["claude-code-telegrammer>=0.2.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "websockets>=12.0",
    "aiosqlite>=0.20",
    "deepeval>=1.0",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=2.0",
    "sphinx-copybutton>=0.5",
    "sphinx-autodoc-typehints>=1.25",
]

[project.scripts]
scitex-orochi = "scitex_orochi._cli:main"
scitex-orochi-server = "scitex_orochi._server:main"
scitex-orochi-mcp = "scitex_orochi.mcp_server:main"

[project.entry-points."scitex_dev.skills"]
scitex_orochi = "scitex_orochi"

[project.entry-points."scitex_dev.docs"]
scitex_orochi = "scitex_orochi"

[project.urls]
Homepage = "https://github.com/ywatanabe1989/scitex-orochi"
Repository = "https://github.com/ywatanabe1989/scitex-orochi"

[tool.hatch.build.targets.wheel]
packages = ["src/scitex_orochi", "hub", "orochi"]
# Explicit data-file inclusion. Hatch ships `.py` by default; everything
# else needs to be listed, or `pip install scitex-orochi` from PyPI ends
# up without the `_skills/` .md files and without the bundled TypeScript
# sidecar. todo#450 surfaced the _skills drift (8 files vs 44 in source)
# — listing `_skills/**` explicitly makes the wheel a faithful copy of
# the package tree for non-editable installs.
artifacts = [
    "_ts/**",
    "src/scitex_orochi/_skills/**",
]

[tool.hatch.build.targets.sdist]
artifacts = [
    "src/scitex_orochi/_ts/**",
    "src/scitex_orochi/_skills/**",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "llm_eval: marks tests that use an LLM-as-judge (DeepEval) and require provider API keys (deselect with '-m \"not llm_eval\"')",
]

[tool.scitex-linter]
library_dirs = ["src", "tests", "hub", "orochi"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
